+++ CMakeLists.txt
@@ -6,7 +6,9 @@ set(CMAKE_BUILD_TYPE Debug)
 add_executable(
     tests
     tests/main.cpp
-    tests/platform_mock_test.cpp  
+    tests/semaphore_mock_test.cpp
+    tests/mutex_mock_test.cpp
+    tests/criticalsection_mock_test.cpp
     mock/semaphores.c
     mock/mutex.c
     )
